Refactor gui.py, remove unnecessary code and add fastapi routing
