+++ imagedephi/gui/__init__.py
@@ -1,0 +1,3 @@
+from .app import app, shutdown_event
+
+__all__ = ["app", "shutdown_event"]
