+++ imagedephi/gui/utils/constants.py
@@ -1,0 +1,1 @@
+MAX_ASSOCIATED_IMAGE_HEIGHT = 160
